move dx dir under src
