--- dx/formatters/__init__.py
@@ -1,3 +1,0 @@
-from .dataresource import *
-from .dx import *
-from .main import *
