+++ src/dx/filtering.py
@@ -123,29 +123,6 @@ def update_display_id(
             metadata=metadata,
         )
 
-    # TODO: replace with custom callout media type or remove altogether
-    # we can't reference a variable type to suggest to users to perform a `df.query()`
-    # type operation since it was never declared in the first place
-    # if not df_name.startswith("unk_dataframe_"):
-    #     output_variable_name = output_variable_name or "new_df"
-    #     # wrapping the triple quotes internally so the user can copy/paste directly
-    #     # without worry of double/single quotes in their data not being handled
-    #     pandas_query_str = f'"""{pandas_filter.format(df_name=df_name)}"""'
-    #     filter_code = (
-    #         f"""{output_variable_name} = {df_name}.query({pandas_query_str}, engine="python")"""
-    #     )
-    #     filter_msg = f"""Copy the following snippet into a cell below to save this subset to a new dataframe:
-    #     <pre style="background-color:white; padding:0.5rem; border-radius:5px;">{html.escape(filter_code, quote=True)}</pre>
-    #     """
-    #     display_callout(
-    #         filter_msg,
-    #         header=False,
-    #         icon="info",
-    #         level="success",
-    #         display_id=display_id + "-primary",
-    #         update=True,
-    #     )
-
 
 def handle_resample(data: dict) -> None:
     # TODO: add resample message to types.py
