+++ src/dx/utils/geometry.py
